+++ examples/app.py
@@ -2,7 +2,7 @@ import streamlit as st
 
 st.set_page_config(layout="wide")
 
-# ------------ Home ------------
+# ------------- Docs -------------
 home = st.Page(
     "./docs/readme.py",
     title="README",
@@ -13,6 +13,11 @@ license = st.Page(
     title="LICENSE",
 )
 
+changelog = st.Page(
+    "./docs/changelog.py",
+    title="CHANGELOG",
+)
+
 # -------- Examples / Demos --------
 node_style = st.Page(
     "./demos/node_style.py",
@@ -21,16 +26,28 @@ node_style = st.Page(
 edge_style = st.Page(
     "./demos/edge_style.py",
     title="Edge Styles",
-)
+)  #TODO: use multi-edge graph example for edge styles
 layout = st.Page(
     "./demos/layout.py",
     title="Layout Algorithms",
 )
 
+#TODO: Return Selection
+
+# --------- Advanced Usage ---------
+event_listeners = st.Page(
+    "./demos/event_listeners.py",
+    title="Events Listeners",
+)
+
+#TODO: Further Styling
+#TODO: Customized Layouts
+
 pg = st.navigation(
     {
-        "Documentation": [home, license],
+        "Documentation": [home, changelog, license],
         "Demos": [node_style, edge_style, layout],
+        "Advanced Usage": [event_listeners],
     }
 )
 pg.run()