+++ st_link_analysis/__init__.py
@@ -1,4 +1,5 @@
 from st_link_analysis.component.component import st_link_analysis
 from st_link_analysis.component.styles import NodeStyle, EdgeStyle
+from st_link_analysis.component.events import Event
 
-__all__ = ["st_link_analysis", "NodeStyle", "EdgeStyle"]
+__all__ = ["st_link_analysis", "NodeStyle", "EdgeStyle", "Event"]